agents/daryxx_finetunned/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.35.0
python-dotenv>=1.0.1
